*** Viewmodel revamp ***
Old viewmodel separated into three parts:

- ChatViewModel.kt
- UserViewModel.kt
- AuthViewModel.kt

*** Repository revamp ***

New repositories based on models instead of data source.

- ChatRepository.kt
- UserRepository.kt
